requirements.txt
git+https://github.com/warvariuc/pyaiml
aiml==0.8.6
aiohttp==1.3.0
appdirs==1.4.3
async-timeout==1.2.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.1.23
chardet==2.3.0
click==6.7
cmake==0.7.1
codegen==1.0
colorama==0.3.7
cycler==0.10.0
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
django==1.11rc1
django-allauth==0.31.0
django-bootstrap-form==3.2.1
django-cors-headers==2.0.2
django-extensions==1.7.8
django-facebook==6.0.3
django-forms-bootstrap==3.1.0
django-nose==1.4.4
django-toolbelt==0.0.1
djangorestframework==3.6.2
emoji==0.4.5
Facebook==0.0
Flask==0.12.2
future==0.16.0
futures==3.1.1
geos==0.2.1
googlemaps==2.4.6
gunicorn==19.7.1
heroku==0.1.4
itsdangerous==0.24
Jinja2==2.9.5
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.2
motionless==1.3.1
multidict==2.1.4
nose==1.3.7
numpy==1.12.1
oauthlib==2.0.2
olefile==0.44
opencv==1.0.1
pandas==0.20.1
patterns==0.3
Pillow==4.2.1
polyline==1.3.2
psycopg2==2.7.1
pyexcel==0.4.5
pyexcel-io==0.3.3
PyJWT==1.4.2
pynmea==0.6.0
pyparsing==2.2.0
pyserial==3.3
pyTelegramBotAPI==2.3.1
python-dateutil==2.6.0
python-decouple==3.0
python-for-android==0.4
python-telegram-bot==5.3.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
scikit-learn==0.18.2
sh==1.12.11
six==1.10.0
sklearn==0.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
static3==0.7.0
staticmap==0.5.2
tabulate==0.7.7
telebot==0.0.3
telepot==10.5
tests==0.7
texttable==0.8.8
Unidecode==0.4.20
urllib3==1.20
virtualenv==15.1.0
Werkzeug==0.12.2
whitenoise==3.3.0
yarl==0.10.0